Moved handleNewItem to App Moved handleSubmit and Form to DataCards Deleted contents of Interact
