+++ our-project/src/components/App.js
@@ -1,11 +1,22 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { Switch, Route } from 'react-router-dom'
 import Home from "./Home"
 import NavBar from "./NavBar"
+import Card from './Card'
 import DataCards from "./DataCards"
 import Interact from "./Interact"
 
 function App() {
+
+
+  const [coffees, setCoffees] = useState([])
+
+  // Create a new array, add the new item, then sets the coffees obj
+  function handleAddItem(newItem) {
+    const updatedCards = [newItem, ...coffees]
+    setCoffees(updatedCards)
+  }
+
   return (
     <div>
       <NavBar />
@@ -17,7 +28,7 @@ function App() {
           <DataCards />
         </Route>
         <Route path="/interact">
-          <Interact />
+          <Interact handleAddItem={handleAddItem} />
         </Route>
         
       </Switch>
